UAST: utilize resolve everything test base
